fix(#88): split auth api from user api
