+++ apps/backend/core/src/main/java/com/schemafy/core/user/controller/UserController.java
@@ -1,72 +1,37 @@
 package com.schemafy.core.user.controller;
 
 import com.schemafy.core.common.constant.ApiPath;
-import com.schemafy.core.common.exception.BusinessException;
-import com.schemafy.core.common.exception.ErrorCode;
-import com.schemafy.core.common.security.jwt.JwtTokenIssuer;
 import com.schemafy.core.common.type.BaseResponse;
-import com.schemafy.core.user.controller.dto.request.LoginRequest;
-import com.schemafy.core.user.controller.dto.request.SignUpRequest;
 import com.schemafy.core.user.controller.dto.response.UserInfoResponse;
 import com.schemafy.core.user.service.UserService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
 @Slf4j
 @RestController
-@RequestMapping(ApiPath.API)
+@RequestMapping(ApiPath.AUTH_API)
 @RequiredArgsConstructor
 public class UserController {
     private final UserService userService;
-    private final JwtTokenIssuer jwtTokenIssuer;
 
-    @PostMapping("/users/signup")
-    public Mono<ResponseEntity<BaseResponse<UserInfoResponse>>> signUp(
-            @Valid @RequestBody SignUpRequest request) {
-        return userService.signUp(request.toCommand())
-                .map(user -> jwtTokenIssuer.issueTokens(
-                        user.getId(),
-                        BaseResponse.success(UserInfoResponse.from(user))));
-    }
-
-    @PostMapping("/users/login")
-    public Mono<ResponseEntity<BaseResponse<UserInfoResponse>>> login(
-            @Valid @RequestBody LoginRequest request) {
-        return userService.login(request.toCommand())
-                .map(user -> jwtTokenIssuer.issueTokens(
-                        user.getId(),
-                        BaseResponse.success(UserInfoResponse.from(user))));
-    }
-
-    @PostMapping("/users/refresh")
-    public Mono<ResponseEntity<BaseResponse<Void>>> refresh(
-            ServerHttpRequest request) {
-        return Mono.fromCallable(() -> extractRefreshTokenFromCookie(request))
-                .flatMap(userService::getUserIdFromRefreshToken)
-                .map(userId -> jwtTokenIssuer.issueTokens(userId,
-                        BaseResponse.success(null)));
-    }
-
-    private String extractRefreshTokenFromCookie(ServerHttpRequest request) {
-        var refreshTokenCookie = request.getCookies().getFirst("refreshToken");
-        if (refreshTokenCookie == null) {
-            throw new BusinessException(ErrorCode.MISSING_REFRESH_TOKEN);
-        }
-        return refreshTokenCookie.getValue();
-    }
-
+    /**
+     * 인증된 사용자 본인의 정보만 조회 가능합니다.
+     *
+     * @param userId 조회할 사용자 ID
+     * @return 사용자 정보
+     */
     @GetMapping("/users/{userId}")
     @PreAuthorize("authentication.name == #userId")
     public Mono<ResponseEntity<BaseResponse<UserInfoResponse>>> getUser(
             @PathVariable String userId) {
-        return userService.getUserById(userId)
-                .map(BaseResponse::success)
+        return userService.getUserById(userId).map(BaseResponse::success)
                 .map(ResponseEntity::ok);
     }
 }
